Move 'NSPersistentContainer' and 'NSManagedObjectModel' helpers to a separate file
